Start of setting up the display limit
This is to prevent display overrun, however on creation of this branch a new bug was discovered.  Should be able to select a new operator after typing in text and continue with the calculation using the new operator.  This isn't the case, so will be chasing down this bug first prior to working on the display issue.